+++ generate-mesh/2d/symmetric-square/change_out.sh
